fix(backend): extract custom field cache into a lightweight module to break a circular import
